requirements.txt
fastapi
uvicorn[standard]
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

# additional dependencies for pose detection and analysis
opencv-python
mediapipe
numpy
pynput
python-multipart
razorpay
